Periodic polling and caching in knitgateway (#10)
This change enables periodic polling of the schema as well
as support for caching the schema. This allows the schema to be
live-reloaded as it changes without restarting the gateway. The caching
allows for the gateway to work even if a BSR or gRPC reflection backend
is unreachable at startup (when the schemas are first downloaded).