Add admin endpoints to knitgateway (#24)
These include "reload now" endpoint, so in dev environments,
there is a way to trigger reloading of the schema. There are also
endpoints for viewing the config that the server was started with
and also viewing some of the current internal state for all of the
supported methods and their associated schemas.