Fix goreleaser config for setting version number (#47)
This isn't a real issue in practice since the default value of the var
should be correct. This configuration for `goreleaser` is only in place
in the off-chance that we tag a sha where the var has not been correctly
updated. In that case, both `make install` and `goreleaser` will update
the var so the built binary emits the correct version.